Move actor proxies creation methods to an IActorProxyFactory extension class
