+++ src/Bullfrog.Actors/ActorProxyFactoryExtensions.cs
@@ -1,0 +1,24 @@
+﻿using Bullfrog.Actors.Interfaces;
+using Microsoft.ServiceFabric.Actors;
+using Microsoft.ServiceFabric.Actors.Client;
+
+namespace Bullfrog.Actors
+{
+    internal static class ActorProxyFactoryExtensions
+    {
+        public static IScaleEventStateReporter GetScaleEventStateReporter(this IActorProxyFactory proxyFactory, string scaleGroup)
+        {
+            return proxyFactory.CreateActorProxy<IScaleEventStateReporter>(new ActorId("reporter:" + scaleGroup));
+        }
+
+        public static TActor GetActor<TActor>(this IActorProxyFactory proxyFactory, string scaleGroup, string region)
+               where TActor : IActor
+        {
+            var actorName = typeof(TActor).Name;
+            if (actorName.StartsWith('I'))
+                actorName = actorName.Substring(1);
+            var actorId = new ActorId($"{actorName}:{scaleGroup}/{region}");
+            return proxyFactory.CreateActorProxy<TActor>(actorId);
+        }
+    }
+}
