+++ src/Bullfrog.Actors/BullfrogActorBase.cs
@@ -1,10 +1,8 @@
 ﻿using System.Threading.Tasks;
 using Bullfrog.Actors.EventModels;
-using Bullfrog.Actors.Interfaces;
 using Bullfrog.Common.Telemetry;
 using Eshopworld.Core;
 using Microsoft.ServiceFabric.Actors;
-using Microsoft.ServiceFabric.Actors.Client;
 using Microsoft.ServiceFabric.Actors.Runtime;
 
 namespace Bullfrog.Actors
@@ -15,15 +13,13 @@ namespace Bullfrog.Actors
     public abstract class BullfrogActorBase : Actor
     {
         private ActorMethodDuration _actorMethodDurationEvent;
-        private readonly IActorProxyFactory _proxyFactory;
 
         protected IBigBrother BigBrother { get; }
 
-        protected BullfrogActorBase(ActorService actorService, ActorId actorId, IBigBrother bigBrother, IActorProxyFactory proxyFactory)
+        protected BullfrogActorBase(ActorService actorService, ActorId actorId, IBigBrother bigBrother)
             : base(actorService, actorId)
         {
             BigBrother = bigBrother;
-            _proxyFactory = proxyFactory;
         }
 
         protected override Task OnPreActorMethodAsync(ActorMethodContext actorMethodContext)
@@ -42,21 +38,6 @@ namespace Bullfrog.Actors
             BigBrother.Publish(_actorMethodDurationEvent);
             _actorMethodDurationEvent = null;
             return base.OnPostActorMethodAsync(actorMethodContext);
-        }
-
-        protected IScaleEventStateReporter GetScaleEventStateReporter(string scaleGroup)
-        {
-            return _proxyFactory.CreateActorProxy<IScaleEventStateReporter>(new ActorId("reporter:" + scaleGroup));
-        }
-
-        protected TActor GetActor<TActor>(string scaleGroup, string region)
-               where TActor : IActor
-        {
-            var actorName = typeof(TActor).Name;
-            if (actorName.StartsWith('I'))
-                actorName = actorName.Substring(1);
-            var actorId = new ActorId($"{actorName}:{scaleGroup}/{region}");
-            return _proxyFactory.CreateActorProxy<TActor>(actorId);
         }
     }
 }
